% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilmoregirls-package.R
\docType{package}
\name{gilmoregirls-package}
\alias{gilmoregirls}
\alias{gilmoregirls-package}
\title{gilmoregirls: The Entire Transcript from Gilmore Girls in Tidy Format}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

"So... it's a show? It's a lifestyle. It's a religion." The complete scripts from the American TV Show Gilmore Girls. Use this package to analyze and practice data wrangling and text.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lgiselebock/gilmoregirls}
  \item Report bugs at \url{https://github.com/lgiselebock/gilmoregirls/issues}
}

}
\author{
\strong{Maintainer}: Luisa Böck \email{luisagisele@gmail.com}

}
\keyword{internal}
