% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilmoregirls_info.R
\docType{data}
\name{gilmoregirls_info}
\alias{gilmoregirls_info}
\title{Episode Information}
\format{
A \link[tibble:tibble-package]{tibble} with 153
rows and 11 variables:
\code{index}, \code{season}, \code{episode}, \code{title},
\code{directed_by}, \code{written_by}, \code{air_date},
\code{us_views_millions}, \code{imdb_rating} and \code{description},
\code{collected_date}.
}
\source{
\url{https://pt.wikipedia.org/wiki/Gilmore_Girls}

\url{https://www.imdb.com/title/tt0238784/}
}
\usage{
gilmoregirls_info
}
\description{
This \link[tibble:tibble-package]{tibble} contains information about each of
the episodes. Information was sourced from Wikipedia and IMDb on
June 26th, 2022.
}
\keyword{datasets}
